CustomAnnotations2

+++ src/main/java/com/SampleProjSpring/Demo/annotations/EmployeeAgeValidator.java
@@ -3,9 +3,9 @@ package com.SampleProjSpring.Demo.annotations;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
-public class EmployeeAgeValidator implements ConstraintValidator<EmployeeAgeValidation, Long> {
+public class EmployeeAgeValidator implements ConstraintValidator<EmployeeAgeValidation, Integer> {
     @Override
-    public boolean isValid(Long inputAge, ConstraintValidatorContext constraintValidatorContext) {
+    public boolean isValid(Integer inputAge, ConstraintValidatorContext constraintValidatorContext) {
 
         for(int i=2;i<inputAge;i++){
             if(inputAge % i == 0){
